Working on converting the upgrade script into the CMS
Added a serpent API model, this will return all Modules and Themes as well as manage core updates.

Added make_nodeload() that converts Github download links into nodeload links ( skips the Redirect from Github ) and allows the files to be downloaded using Curl.

Added is_update() this is a place holder at the moment but will server as a flag to prompt for updates.